requirements.txt
pandas
xlrd
rdflib
python-slugify